+++ animations/src/components/StoreList.js
